pyproject.toml
[project]
name = "extractly"
version = "0.1.0"
description = "Extractly is a Streamlit app for defining document schemas, classifying incoming files, and extracting structured metadata with traceable runs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.0",
    "boto3>=1.38.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "ruff>=0.12.3",
    "streamlit>=1.45.1",
]

[dependency-groups]
dev = [
    "commitizen>=4.7.1",
    "mkdocs>=1.6.1",
    "mkdocs-markdownextradata-plugin>=0.2.6",
    "mkdocs-material>=9.6.14",
    "mkdocs-material-extensions>=1.3.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.10",
]

[tool.setuptools]
package-dir = {"" = "src"}   # tells setuptools that code lives in /src
